Handle address mappings with non-zero offset base
This patch updates AddressMapper to support the case where we map a
middle section of a DSO, rather than from the beginning.

e.g. Suppose you have a DSO at 0x10000 with size 0x8000, but only want
to map the range [0x12000, 0x13000). The address 0x12800 would thus map
to an offset of 0x2800.

The old system did not support such a mapping, because there is no
concept of a base offset (analogous to pgoff in kernel/perf). If we
simply mapped{ addr=0x12000, size=0x1000 }, the address 0x12800 would
map to an incorrect offset of 0x800.

In this patch, we allow the mapping to happen with the following
parameters:
  addr=0x12000, size=0x1000, offset_base=0x2000

The parameter |offset_base| means that |addr| is actually offset from
the start of the DSO by |offset_base|. Hence, the address 0x12800 will
map to 0x2800 insetead of 0x800.

This behavior typically happens when something like a VDSO or //anon is
mapped into the middle of an existing DSO. The existing DSO gets split
into two remaining parts before and after the new DSO. The latter part
requires a nonzero offset base.

However, no changes have been required in the test data. It's likely
that we were not getting any samples in the latter remaining part in
such a scenario, so the problem never manifested itself as incorrect
data.

BUG=chromium:372024
TEST=unit tests pass

Change-Id: Ie18728834cb4cb48d536e67b297147e8a8b48301
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202517
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David Sharp <dhsharp@chromium.org>