common-mk: clean up logic to work for qemu & native runs
The current native tests are run in the host sdk env and just set
LD_LIBRARY_PATH to the sysroot.  This means we end up using the
wrong C library and we might pick up wrong third party libs.  Or
if the unittests fork programs which have a scrubbed environment.

The qemu logic isn't much better.  It also doesn't really work if
you want to run in parallel, or if you fork programs (as those won't
be run through qemu).

If a test failed, it had the possibility of leaking bind mounts in
the sysroot (depending on when the user hit CTRL+C).  Not a common
case, but still a possibility, and we don't like that.

Let's gut the whole test logic and take some tips from the common.mk
code.  Specifically, we now do:
 - Always initially `sudo` and run the python code as root.  This
   isn't great, but we need this for later test setups.
 - Always run in our own mount namespace.  This guarantees that the
   bind mounts we create won't leak -- as soon as we exit, they are
   removed by the kernel automatically.
 - Move the bind mount logic from pre_test to run.  This is because
   any mounts we create in the pre_test process will be lost.  It
   means slightly more overhead, but it also means we can now cleanly
   run in parallel.
 - Always run tests in the sysroot chroot (not just for arm/qemu).
   This lets us drop the LD_LIBRARY_PATH hack and guarantees that
   all files read and programs run (fork/system) come only from the
   sysroot.  It also partially fixes the qemu system case.
 - Register qemu as a binfmt handler.  This way we can run target
   binaries as if they were native.  This means we can fully support
   the fork/exec scenario when using qemu.

The other nice thing about this logic is that we have just one code
path now for both the ldso & qemu cases.  The only thing we do in
the qemu case is register a handler, otherwise the rest is the same.

BUG=chromium:401876
TEST=`FEATURES=test emerge-x86-alex platform2` passes
TEST=`FEATURES=test emerge-daisy platform2` passes
TEST=`FEATURES=test emerge-lumpy platform2` passes
TEST=`cbuildbot {amd64,arm,x86}-generic-full` passes

Change-Id: I7003b681fc1008fe14ab468832698e7ce49f3898
Reviewed-on: https://chromium-review.googlesource.com/211054
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>