glbench: remove unit from test name
The test unit instructs graphics_GLBench how to interpret the score, but
instead of prepending it to the testname [and image name], we can just
print it once on the score line, and let graphics_GLBench parse it from
there.

A later patch also removes the unit from the image name, but that is an
independent chage that involves renaming a ton of files, so we leave it
to separate patch.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=test_that graphics_GLBench

Change-Id: Ia9d95b51e2b7297a907c1b52972fb04ac4815273
Reviewed-on: https://chromium-review.googlesource.com/202887
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>