shill: test-scripts: Ignore InProgress error for EnableTechnology.
FlimFlam.EnableTechnology() is (only) used by network3g tests to enable
cellular technology, which in turns enables the cellular modem in the
DUT. If an enable operation is already in progress, EnableTechnology()
could fail with an 'org.chromium.flimflam.Error.InProgress' error. This
is not an issue, and thus shouldn't be considered as an error, in
network3g tests, which always poll for the device state until the device
is enabled after calling EnableTechnology().

BUG=chromium:385413
TEST=Tested the following:
1. network_3GRecoverFromGobiDesync.gobi2k on mario
2. network_3GRecoverFromGobiDesync.gobi3k on alex, daisy

Change-Id: I5a5984eb92eb416db67139be4f4d0fa72259a693
Reviewed-on: https://chromium-review.googlesource.com/204625
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>