chromium_command_builder: use "neon" USE flag rather than "exynos"
libvpx (libyuv, etc.) cannot read cpuinfo to do runtime neon
detection because of chrome's sandbox (crbug.com/161834).
If cpuinfo is not available, libvpx instead uses neon if the env
variable VPX_SIMD_CAPS is 0xf.

Therefore, for chromium os systems that support neon we would like to
set that env variable.

Naming it "exynos" is confusing, since neon is present on many arm
boards, not all of which use samsung exynos SoCs.

A better name is "neon" since that is really the feature we are
detecting.  As an added bonus, all arm boards that support neon
already define USE="neon" in their make.conf, so we don't need to
set a USE flag anywhere for this change.

In theory this USE flag should also be used to set any other flags /
env variables needed by chrome and its third_party libraries that
want neon detection.  For example, this could be used to set an env
variable for libyuv to tell it to use neon instead of hard-coded it
to always use neon on arm systems.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:161834
TEST=cbuildbot:
   daisy-full arm-generic-full nyan-full link-full peach_pi-full

Change-Id: Ida3af1f549d73b6c4a109c0453f001dcd5d21a26
Reviewed-on: https://chromium-review.googlesource.com/204815
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Thomas Inskip <tinskip@google.com>