shill: Treat more modemmanager errors as wrong APN errors.
The network can reject a connection attempt with an incorrect APN with a number
of different NwError error codes that get bubbled up to us through
modemmanager. This CL adds some more error codes to those that we treat as an
incorrect APN error.

BUG=chrome-os-partner:28675
TEST=Attempt to connect to T-Mobile and AT&T network with the APN try list
     changed such that we try incorrect APN(s) first. shill should retry and
     eventually connect using the correct APN.

Change-Id: Ia09bc2ff12d297fd0f2b02698c92dbf3775a89b9
Reviewed-on: https://chromium-review.googlesource.com/203878
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>