common.mk: Disable C++ exceptions by default.
This patch adds a new CXXEXCEPTIONS flag to common.mk that controls
whether the C++ exceptions are allowed. The setting disables the C++
exception support by default.

A minor update to the BASE_VER on login_manager is included here to
support building it outside the ebuild.

BUG=chromium:386443
TEST=emerge-link chromeos-login works (login_manager uses this common.mk).
TEST=VERBOSE=1 emerge-link chromeos-login shows the -fno-exceptions flag.
TEST=cbuildbot {x86,arm}-generic-full

Change-Id: I7e9d64aa6250d6c71562ec831484beae3b2c620d
Reviewed-on: https://chromium-review.googlesource.com/205331
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>