shill: Service: Remove DHCP minimal config code
This code does not end up being very useful (as evidenced by
UMA stats)  and can reduce functionality in infrastructures
that require the use of some DHCP options.

This reverts commit 26a997b04bd65a6252987ea6fdcbbb70785d1d8d,
except for the cleanup of "*service" in unit tests and the
setting of SelectedService in ethernet.cc.

BUG=chromium:360452
TEST=Unit tests

Change-Id: Ib2c49fc5639452bfffd5cd03c1bb69e32568cc60
Reviewed-on: https://chromium-review.googlesource.com/205588
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>