UpdateManager: UpdateCurrentConnectionAllowed should return kAskMeAgainLater.
This reflects the fact that the restriction is temporary, and that the
update process should not be terminated, therefore is more correct than
returning kFailed.

This also relocates the code that allows user settings to enable update
over a cellular connection, in cases where a policy is active but does
not specify allowed connections.

BUG=chromium:390264
TEST=Unit tests.

Change-Id: Iadfd05ba3c4617a22c69b82232b7802d1ea65504
Reviewed-on: https://chromium-review.googlesource.com/206144
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>