login_manager: ui: Create cgroup freezer for chrome renderers
In order to keep power consumption low for lucid sleep, we need to
freeze the chrome renderers.  Create this cgroup in the ui job after
the cpu cgroup and rename the cpu cgroup since it is no longer the only
one.

BUG=chromium:364339
TEST=manual
Deployed on peppy and verified that the chrome_renderers freezer cgroup
was created at startup

Change-Id: Iacf3796c275f374e40c94a54c4295410ee9b2029
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207024
Reviewed-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>