capture_utility: support WP2
The capture_utility script would fail on WP2 systems, because
WP2 rejects attempts to configure the channel (with EBUSY).

For the "monitor-connection-on" functionality, and assuming
there's only one WiFi adapter on the system, we don't actually
need to change the channel. So, in that case, if configuring
the channel fails, fall back to just using or creating any
monitor device on the same phy as the user-specified device.

Also, make sure we configure a couple more things that are
necessary for good pcaps on WP2. First, disable power-save
mode. Second, disable beacon filtering.

While there:
- fix parsing of ifconfig output (strip ':' after device name)
- fix some missing local declarations
- fix a whitespace bug
- fix a typo ('inocuous')
- tweak documentation for get_monitor_for_link

BUG=chromium:392241
TEST=capture_utility_test, manual (see below)

Manual test
-----------
- on WP2 device:
  1. connect to a wifi network
  2. capture_utility.sh --monitor-connection-on wlan0 --output-file /tmp/foo
  3. <ctrl-c>
  4. tcpdump -n -r /tmp/foo
     -> check for sane output
  - repeat steps 1-4 one more time
    (the second time tests what happens when an existing monitor device exists)
- on ath9k device:
  1. connect to a wifi network
  2. capture_utility.sh --monitor-connection-on wlan0 --output-file /tmp/foo
  3. <ctrl-c>
  4. tcpdump -n -r /tmp/foo
     -> check for sane output
  - repeat steps 1-4 one more time
    (the second time tests what happens when an existing monitor device exists)
- on either device:
  1. capture_utility.sh --device eth0 --output-file /tmp/foo
  -> ensure this outputs "Capturing from eth0.", rather than an error

Change-Id: I17e80241c367d7a2d308c2228403f8274ba60c97
Reviewed-on: https://chromium-review.googlesource.com/207095
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>