shill: Distinguish errors when failed to register a DBus object path.
CL:207344 introduces a DBus::ErrorObjectPathInUse error in dbus-c++ to
help identify the type of error when it fails to register a DBus object
path. This CL makes the corresponding changes in DBusControl to
distinguish different types of error when failed to register a DBus
object path.

BUG=chromium:392399
TEST=Tested the following:
1. `FEATURES=test emerge-$BOARD platform2`
2. Manually modify shill to register a DBus adaptor with the same object
   path twice, and verify that it logs "object path in use".

Change-Id: I90d895ef162f72be68077c7f6562926bccaff588
Reviewed-on: https://chromium-review.googlesource.com/207465
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>