shill: cellular: Add a flag to choose from duplicate provider entries.
serviceproviders.prototxt has some MobileNetworkOperator entries that have the
same MCCMNC/operator name. Although the right thing is to fix this duplication
(some entries may be obsolete, some entries may actually be MVNO), it can
sometimes take a while to fix this since it involves interacting with the
providers.

As a stop gap solution whenever such a duplication causes shill to not recognize
an operator, this CL adds a bool flag |earmarked| to the database to pick out
the operator that must be chosen among the duplicates.

BUG=chromium:390968
TEST=(1) Run unittests.

Change-Id: If3724be62c193993ea39b7facf92bc297827fef3
Reviewed-on: https://chromium-review.googlesource.com/208465
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>