shill: wifi: improve effectiveness of mac80211_monitor
If we're asked to disconnect, when not connected, immediately
force ourselves into the disconnected state. We can't rely on
supplicant to emit a change notification for CurrentBSS in this
case.

The problem seems to be due to the behavior of sme_authenticate().
When wpa_supplicant roams from on BSS to another, it clears its
notion of the current bssid:

    os_memset(wpa_s->bssid, 0, ETH_ALEN)

It then emits a change notification for the current BSS

    if (bssid_changed)
      wpas_notify_bssid_changed(wpa_s)

This leads (eventually) to wpas_dbus_getter_current_bss, which
references wpa_s->current_bss->id. This is NOT wpa_s->bssid.
Instead, wpa_s->current_bss is set, by sme_authenticate(), to
be the BSS that supplicant is roaming to.

That keeps shill happy, because shill believes that supplicant
is still connected or connecting.

But, if the authenticate attempt fails (e.g. it times out),
then we're in for trouble. The reason is that bssid_changed is
computed as follows:

  bssid_changed = !is_zero_ether_addr(wpa_s->bssid)

This logic appears in sme_authenticate(), and many other places.
So, once supplicant hits sme_authenticate(), shill might not be
informed of BSS changes. In particular, shill might be told that
supplicant has given up, and set its current BSS to all zeros.

It appears that, in the normal case, wpa_s->bssid is set to the
an actual (non-zero) BSS after association completes.

We work around this by forcing ourselves into the disconnected
state on a DisconnectFrom call, unless wpa_supplicant reports
that the interface state is Completed.

This changes improves the efficacy of Mac80211Monitor from
about 70%, to about 95%. (Each of these is measured on at least 250
trials. The Mac80211Monitor is considered effective if it gets us
back online with 90 seconds of the TX queues being stopped.)

I haven't examined the remaining 5% in depth. In at least one case,
however, the problem was that we'd blacklisted all of the APs with
decent signal. So the attempt to reconnect failed because of poor
signal, rather than being stuck.

BUG=chromium:392209
TEST=manual

Manual testing: verified per procedure in crbug.com/395234#c7.
Note that the testing requires CL:208980, which is not present in
the tree.

Change-Id: Iee6a173b23da695ef658e7bdb210ec3281e5e9dc
Reviewed-on: https://chromium-review.googlesource.com/210450
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>