powerd: Use pm_test to recover from dark resume
When we are in dark resume, some devices are powered on while others are
powered off.  Once selective resume is ready, we should be able to go
directly from this state to a fully powered on state.  Until then, we
have to (ab)use the pm_test mechanism to get all the devices to run
their resume code.

Additionally, since we will not be backporting selective resume to older
kernels but we still plan on enabling dark resume for most devices that
are using those kernels, ensure that all user activity is ignored while
in dark resume on devices with those kernels.

BUG=chromium:397346,chromium:397777
CQ-DEPEND=CL:211519
TEST=unit tests and manual
Tested on link with kernel-3.14 and "0.0 10" in
dark_resume_suspend_durations.  Poked the system when it appeared to
come up in dark resume and then checked the logs to verify that the
transition message appeared.

Tested on link with kernel-3.8 and "0.0 10" in
dark_resume_suspend_durations.  Poked the system when it appeared to
come up in dark resume and saw that the system did not resume.  Then I
triggered a regular resume and checked the logs to verify that there was
no transition message.

Change-Id: I74f69bfc8a94e3e31574f60e54efb8dea91f7133
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209963