powerd: Set up dark resume behavior for lucid sleep
This sets up all the implementation-agnostic code that we will need for
lucid sleep.  The actual behavior doesn't change with the exception that
we now read the wakeup count even in a dark resume.  The reasoning for
this is that we're going to be spending a lot more time in dark resume,
increasing the chances of a user interacting with the system in the
middle of one.  If we want the kernel to cancel the suspend when this
happens, we need to provide it an accurate wakeup count.

BUG=chromium:397346
TEST=unit tests
CQ-DEPEND=CL:209962

Change-Id: I08f850eddf97aabbba63c31d9a803a2304a0aefc
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209964