shill: cellular: Prefer operator name from operator database.
Before this CL, the operator name provided by the SIM or obtained OTA was given
preference over that obtained from the mobile network operator database.
This CL changes the preference so that the database information is treated as
the ground truth.

Since information other than the operator name can be used to key-in on an MVNO,
the MVNO information from the database should be able to override the operator
name obtained from the modem.

BUG=chrome-os-partner:23086
TEST=(1) Run shill unittests.
     (2) Test MVNO identification in field.

Change-Id: I5111281a84427b0e6ec28e2decfa04c2e47ee93a
Reviewed-on: https://chromium-review.googlesource.com/210079
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>