init: Move init/test/uinput.conf to init/
The uinput kernel module is required by Bluetooth daemon to provide the
AVRCP profile. This configuration file is responsible for loading the
module and by moving it to the init/ directory, we make it available to
all images, not only test.

This file is currently installed by chromeos-test-init package. After
this patch is merged, it will then be installed by chromeos-init.

BUG=chromium:256746
R=keybuk
TEST=build base or dev images and check for /etc/init/uinput.conf file

Change-Id: I42b0c8761a637f1401aaf81faec126de2211a47a
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/211034
Reviewed-by: Nathan D Ciobanu <nathan.d.ciobanu@intel.com>
Tested-by: Nathan D Ciobanu <nathan.d.ciobanu@intel.com>
Reviewed-by: Kees Cook <keescook@chromium.org>